Change default syntax highlighting theme
The default syntax highlighting theme has been changed to "default", as it works best on both light and dark terminals.

I have also added a settings menu with the ability to change the syntax highlighting theme. This change is still not persisted yet, but will be written to a config file eventually.