Add `--run` command line flag to run notebooks.
If enabled, notebooks are run automatically when they are opened. If used with `--dump`, the notebook is run before cells are printed.

Also includes a major re-factoring of the notebook's kernel, separating it into it's own class. It also adds the ability to interrupt the kernel, and fixes kernel changes.