Implement new terminal graphics system.
This fixes issues relating to displaying images with kitty graphics, and only
unrenders graphics to ascii images if the image itself is obscured, rather than
the entire cell.

This paves the road to cropping images and displaying only part of them if they
are partially out of view.

This also refactors the render system.