Implement a new terminal feature detection system
This uses prompt-toolkit's input parsing system to detect terminal responses:
expected terminal response sequences are assigned a `key` (much like like a
key-press), and a callback can be triggered when a response is recieved.

In order to detect the terminal's responses, rendering is paused when the app
is launched (but not input processing), to allow for responses from the
terminal (which come back via stdin).