Implement new smoother scrolling system.
This re-write makes rendering more efficient, reducing lags and flickering when
rendering a notebooks.

This also required a new refactoring of the terminal graphics system, which is
now much improved: graphics are shown without delay when opening a notebook,
and hide when a menu or dialog is shown to prevent breaking the display.

Also adds key-bindings to dialogs.

Also causes cells to run interactively when run all is called in TUI mode.

Also fixes some typos in comments.